boot: generalize boot_write_image_ok() function to get flag as parameter
This commit generalizes boot_write_image_ok() function to take flag
as an input parameter. Function is also rename to boot_write_image_flag()
so the name matches the usage better.

This is useful to future implementation of different algorithms that
might need other flags than BOOT_FLAG_SET to be written to image
trailer.

Signed-off-by: Michal Lenc <michallenc@seznam.cz>